+++ Moyoy-Api/src/main/java/com/moyoy/api/pr_review/application/PrReviewService.java
@@ -4,9 +4,7 @@ import java.time.LocalDateTime;
 
 import lombok.RequiredArgsConstructor;
 
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.moyoy.api.pr_review.application.request.PrReviewCreateData;
 import com.moyoy.api.pr_review.application.request.PrReviewUpdateData;
@@ -14,10 +12,8 @@ import com.moyoy.api.pr_review.application.request.SearchConditionData;
 import com.moyoy.api.pr_review.application.response.*;
 
 import com.moyoy.domain.pr_review.PrReview;
-import com.moyoy.domain.pr_review.PrReviewHit;
 import com.moyoy.domain.pr_review.PrReviewHitRepository;
 import com.moyoy.domain.pr_review.PrReviewRepository;
-import com.moyoy.domain.pr_review.dto.PrReviewHitCreate;
 import com.moyoy.domain.pr_review.error.PrReviewDeleteForbiddenException;
 import com.moyoy.domain.pr_review.error.PrReviewEditForbiddenException;
 import com.moyoy.domain.pr_review.error.PrReviewNotFoundException;
@@ -32,10 +28,10 @@ import com.moyoy.common.page.SliceResult;
 @RequiredArgsConstructor
 public class PrReviewService {
 
-	private final PrReviewService self;
+	private final PrReviewAsyncService prReviewAsyncService;
+
 	private final PrReviewRepository prReviewRepository;
 	private final PrReviewQueryRepository prReviewQueryRepository;
-	private final PrReviewHitRepository prReviewHitRepository;
 
 	public PrReviewListResult getPrReviewList(SearchConditionData condition) {
 
@@ -53,7 +49,7 @@ public class PrReviewService {
 
 		/// TODO: 조회수 관리 (v1: 브릿지 테이블, v2: Redis, v3: Redis 기록/누적조회 일괄 업데이트, v4: hyperloglog/누적조회 일괄 업데이트)
 		///  특히 조회수 증가는 도메인 서비스인데, 뷰목적 dto 뿐아니라 entity로 도메인도 가져와야함.
-		self.increaseHitAsync(reviewId, userId);
+		prReviewAsyncService.increaseHitAsync(reviewId, userId);
 
 		return PrReviewDetailResult.from(data, isWriter);
 	}
@@ -113,23 +109,5 @@ public class PrReviewService {
 		Long closedReviewId = prReviewRepository.save(prReview).getId();
 
 		return new PrReviewCloseResult(closedReviewId);
-	}
-
-	@Async("hitsExecutor")
-	@Transactional
-	public void increaseHitAsync(Long reviewId, Long userId) {
-		/// 해당 트랜잭션이 실패하면, 후처리를 할 것인가.
-
-		LocalDateTime now = LocalDateTime.now();
-
-		PrReviewHit hit = prReviewHitRepository.findOrCreate(
-			PrReviewHit.create(new PrReviewHitCreate(reviewId, userId, now)));
-
-		if (!hit.canIncrease(now))
-			return;
-
-		prReviewRepository.increaseHitCount(reviewId);
-
-		prReviewHitRepository.updateLastIncreasedAt(hit.getId(), now);
 	}
 }
